requirements.txt
alembic==1.12.0
asgiref==3.7.2
black==23.9.1
click==8.1.7
coverage==7.3.1
Django==4.2.5
django-filter==23.2
django-redis==5.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
protobuf==4.21.12
pytest==7.4.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
